fix: repair all remaining unbuildable TUs (32 + 8 cascading) — full tree compiles
Redeclaration/type-conflict fixes, verified against retail asm:

- CTTask: drop harness_catalog.hpp where unused (9 TUs incl. CExchangeWin,
  CTaskREvent, code_8018F8D8, CfObjectModel, CMenu*); monolib vs
  CTaskGameEff.hpp template collision resolved
- func_80157C4C: header 1-arg decl wrong -> TU-local in CItemBoxInfo.cpp
  (retail passes index + s16 value); CEquipChange builds
- func_800FF778/getInstance__11CSysWinBuffFv: unify bool/CSysWinBuff*
  returns per retail (thunk + singleton getter stay 100%)
- func_80186BC8/int, func_8004B9B8(+0x14), mBuffer2->_A94: retail-verified
- func_8009D018/func_8009EC9C/func_8008187C/func_800ACF78/func_800ACEF8/
  __dynamic_cast/__ptmf_null: unify signatures across headers (linkage +
  int/u32 + const void* forms); CChainTimer/CHelp_Pg/CModelDispMakeCrystal/
  pluginUnit/pluginBtl/CCharEffectEne/CfGimmickWarp/IResInfo/CfPadTask/
  CtrlMoveNpc build
- nw4r: ResMatFurData local struct dropped (use g3d_resmat.h);
  RemoteSpeakerManager ctor dup removed; SoundArchiveFileReader stub
  signatures fixed
- CREvtCamera lambda -> static helper; func_80086B5C declared
- code_800B06A4: reslist/FixStr layout kept, umbrella include dropped

No FULL_MATCH/EQUIVALENT_MATCH regressions (stash-verified). Registry has
pre-existing stale FULL_MATCH claims from voidtype/pi-harness rewrites
(units never recompiled) - separate recertification task.

+++ src/kyoshin/CArtsInfo.cpp
@@ -290,7 +290,7 @@ void func_80235AE0(CArtsInfo* self) {
     self->field_0x5A = 0;
 
     // Update layout elements with arts data
-    u32 artsData = func_8009EC9C(self->field_0x54);
+    u32 artsData = (u32)func_8009EC9C(self->field_0x54);
     char* str1 = (char*)func_80136190(lbl_eu_8050B00C + 0x32, lbl_eu_8050B00C + 0x3D, 0x18);
 
     char buf[64];
